Move S3ChunkedStreamReader to its own file
